Add arrays and enable binary parsing of messages
ArrayValues can only have elements of the same ScalarValue or of
MessageValues with the same model. Multidimensional arrays are not
supported.

To simplify the parsing of MessageValues from a given bytestring
add the Bitstring class. It helps to perform the assignment
of values to fields with a bitlength which is not a multiple of 8.

Add unit tests with the same functionality as their counterparts in
the SPARK implementation to find missing features.

Ref. #169, #189