Filter special nodes in __computer_topological_sorting
Previously the INITIAL and FINAL nodes were removed from the functions' result
list by simply stripping off the first and last element. However, the
assumption that INITIAL is the first and FINAL is the last element of the list
is not true for invalid models (e.g. those with fields that have no path to
FINAL). As we detect those error cases *after* computing fields,
__compute_topological_sorting should be robust against special fields in
different positions. We simply filter out INITIAL and FINAL which is equivalent
to the previous approach for valid models.

Ref. #100