Improve verification error messages
Instead of large expressions which are manually joined with `And()`, it's
better to feed the facts to z3 using the `add` method. Having the facts
separate also allows us to construct the unsatisfiability core after
proofs failed. This is done by redoing the proof and adding facts using
the `assert_and_track` method. The unsat core is a subset of facts that
leads to the contradiction. While it's not guaranteed to be minimal, it's
usually smaller than the original set of facts which leads to much more
helpful error messages.

Ref. #126