Remove special treatment of refinement identifiers
When using an invalid package name for a refinement (say "A.B"), then
the expected error is supposed to be something like "A.B is not a valid
package name". For refinements, however, the super class is initialized
with

   package * "__REFINEMENT__"
   f"{flat_name(sdu.full_name)}__{flat_name(pdu.full_name)}__{field.name}__"

which produces some convoluted error message mentioning that long
internal string. Instead of filtering such situations in the super-class,
we now reinitialized the error buffer after initializing the super-class
of Refinement, effectively ignoring errors in the ID() constructor. The
well-formedness of the package identifier is checked separately.

Ref. #248, #285