Fix contradiction and field position verification
The fix for #277 was actually not proper, as any contradiction would
be flagged as an error. However, as long as there is at least one
non-contradictory path, we cannot treat a field condition as
contradictory.

This problem occurred in the test, but was not discovered as the test
was missing a error.propagate() call after message construction. This
was noticed when making Message constructors raise exceptions again in
the discussion of #285.

Ref. #248