Provide third party sources to isort
isort 4.3.21 on Debian (installed via pip3) seems to be broken and does
not detect the third party sources. As a consequence, it formats all
sources to remove the blank line between those imports and our local
imports. Specifying them in the config fixes the issue and also makes
isort more robust in cases where the dependencies are not installed.