Enable making all checks required when using matrix
A job using matrix is not expanded into multiple checks when the job is
skipped and thus it is not possible to make these checks required. Each
step of a job must be skipped separately, because GitHub Actions does
not provide a way to skip all subsequent steps of a job.