Correct "empty" line containing space characters
One of the Ada code blocks in 03-getting-started-with-recordflux.rst
contained a "blank" line comprising several space characters.
This resulted in a failure of one of the style checks (performed in
the check_doc job in the push workflow). The errant line has been
trimmed so that it is now truly empty.