Don't depend on known field_size in _field_constraints method
We replace the use of `_target_size` by a similar `_target_size_opt`
function, which returns None if the size of the link target can't be
computed. In this case, no constraints need to be added to the
constraints computed by `_field_constraints`. We do the same for
`_target_last_opt`.

Also add unit test for new `_target_last_opt` function

For eng/recordflux/recordflux#1480