Pass bytes array directly instead of pointer type
In the Insert/Extract options, a pointer to a byte array was passed. We
can change this interface to pass the byte array directly. This
simplifies both these functions as well as their unit tests, and
requires minimal adaptation in the message code. Proof time on
`rflx-builtin_types_tests.adb` is roughly halved.

Ref. eng/recordflux/RecordFlux#1515