Fix code generation for Boolean conditions
In substitutions, field values and parameters are usually converted to
the "target type" (typically Base_Integer). This conversion is now
disabled for Boolean expressions.

We add a simplification that rewrites (dis)equalities of the form "X =
True/False" to simply "X" (or "not X"). The existing simplification
which rewrites expressions such as "True = X" to "X = True" is removed,
as both would now be simplified to "X".

Ref. eng/recordflux/RecordFlux#1365