Prevent unnecessary initialization of language server
As the language server must be a global variable, the language server is
initialized when the module is imported. Thus, the initialization is
done every time the CLI is run and every time the process is forked.
This can be avoided by moving the import into the function that starts
the language server.

Ref. None