Ensure Consistent Python Version Usage
Previously, the native Python module could be built outside the Python
virtual environment. This posed a problem if the virtual environment
used a different Python version than the system default.

If the native module is built for a different version than the
interpreter version, it may cause undefined symbols, as the native
module may attempt to call functions that are not present in the
current interpreter version.

Ref. eng/recordflux/RecordFlux#1747